Set version prefix 'v' in mkdocs.yml for more flexible
Sometimes,
we write codes to auto replace the version with current branch name,
and build a documentation for specific branch or environment.

for example:

- extra.version = master
- extra.version = staging
- extra.version = develop

we will see the version at top bar is `vmaster`, `vstaging` or
`vdevelop`, set a fixed version prefix `v` in HTML template is not flexible,
so just move it into mkdocs.yml from HTML template.